Add user retrieval by email in UsuariosController
Implemented a new action method `UsuarioCorreo` in `UsuariosController.cs` to fetch a user based on their email address. This method is secured with authorization and responds to GET requests.

Added a corresponding method `GetUsuarioCorreo` in `RepositoryCamisetas.cs` to query the database for users by email. Updated the namespace for the repository to `ApiCamisetas.Repositories`.